backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
roboflow==1.1.48
inference-sdk==0.9.22
httpx==0.27.0
python-dotenv==1.0.0
supabase==2.5.0
